GLB export: Fix buffer alignment by adding padding bytes
Previously, it tries to compensate for alignment but it does it too
early. Going "past the buffer" in python just truncates the resulting
array, so no padding bytes are actually added to the output, essentially
causing a desync between the buffer contents and recorded offsets, so
GLB consumers see nonsensical data.

This fixes it by actually adding zero bytes for padding to the output.